Temperature actually changes based on the weather
Oh shit oh fuck.

I have no idea what I'm doing, stop me. This is a bad idea. Basically it grabs the weather holder's temperature and updates the temperature of every turf when the weather changes. I need to test this out for how CPU greedy it is, and if it actually retains the temp.